fix: move client-only widgets to client wrappers
