+++ packages/nextjs/components/TransactionFeed.client.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import dynamic from "next/dynamic";
+
+const TransactionFeed = dynamic(() => import("./TransactionFeed"), {
+  ssr: false,
+  loading: () => null,
+});
+
+export default TransactionFeed;
