+++ packages/nextjs/components/home/HeroSection.tsx
@@ -1,17 +1,7 @@
 import React from "react";
 import Image from "next/image";
 import Link from "next/link";
-import dynamic from "next/dynamic";
-import Spinner from "../common/Spinner";
-
-const DebtComparison = dynamic(() => import("./DebtComparison"), {
-  ssr: false,
-  loading: () => (
-    <div className="flex justify-center py-8">
-      <Spinner size="loading-lg" />
-    </div>
-  ),
-});
+import DebtComparison from "./DebtComparison.client";
 
 const HeroSection = () => {
   return (
